+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [0.4.11] - Unreleased
 ### Added
-* n/a
+* `VectorSearchIndex` Databricks resource
+* `VectorSearchEndpoint` Databricks resource
+* `purge` method for data sink
+* `refresh` option for pipeline and pipeline node
 ### Fixed
 * Checkpoint location of `TableDataSink`
 ### Updated